Break up the protocol file into individual files.
