+++ uma/utils/url_utils.go
@@ -1,4 +1,4 @@
-package uma
+package utils
 
 import "strings"
